Title: 합과 차의 나눗셈, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240325/합과 차의 나눗셈/divide-of-sum-and-sub.cpp
@@ -7,6 +7,6 @@ int main() {
     cin >> a >> b;
     cout << fixed;
     cout.precision(2);
-    cout << (a + b) / (a - b);
+    cout << double(a + b) / double(a - b);
 	return 0;
 }
